% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DJFMeans.R
\name{DJFMeans}
\alias{DJFMeans}
\title{DJFMeans}
\usage{
DJFMeans(ds, flowDir, xVariable, smoothLevel)
}
\arguments{
\item{ds}{The dataset}

\item{flowDir}{The directory of the gated .fcs data}

\item{xVariable}{The fluorescence channel on the x axis}

\item{smoothLevel}{The level of smoothing applied to the data}
}
\description{
A function that will update the peaks identified for the throughout the
analysis for the Dean-Jett-Fox model. The model is sensitive to noisy data
and so we smooth out the data to run the model. DJFMeans will find the peaks
associated with the data at the smoothing level indicated.
}
\examples{
DJFMeans(
ds = flowData,
flowDir = "FlowData/T10_FLC/gated_data",
xVariable = "FITC-A",
smoothLevel = 10
)
}
