% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubletCheck.R
\name{doubletCheck}
\alias{doubletCheck}
\title{doubletCheck}
\usage{
doubletCheck(doubletCheckDs, peaks, g1G2Range, g2G2Range)
}
\arguments{
\item{doubletCheckDs}{The data set to analyze for doublets}

\item{peaks}{The data peaks that have been identified}

\item{g1G2Range}{A range around the G1+G2 location on the x-axis,
i.e identifying the G1+G2 doublet in the range g1G2Range - G1+G2, g1G2Range + G1+G2}

\item{g2G2Range}{A range around the G2+G2 location on the x-axis,
i.e identifying the G2+G2 doublet in the range g2G2Range - G2+G2, g2G2Range + G2+G2}
}
\description{
For single populations, doubletCheck is used to identify doublets.
Doublets are an excess of cells around the G1+G2 range.
}
\examples{
\dontrun{
doubletCheck(
 doubletCheckDs = single_ds,
 peaks = peaks_ds,
 g1G2Range = 10,
 g2G2Range = 15
)
}
}
