% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findClusters.R
\name{findClusters}
\alias{findClusters}
\title{findClusters}
\usage{
findClusters(ds, clusterDist, maxXValue)
}
\arguments{
\item{ds}{The data set to analyze}

\item{clusterDist}{The smallest distance between two peaks/clusters}

\item{maxXValue}{the maximum of the x-axis}
}
\description{
findClusters identifies G1 and G2 sub population pairing.
}
\examples{
findClusters(
 ds = possible_peaks2,
 clusterDist = 40,
 maxXValue = maxXValue
 )

}
