% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPairs.R
\name{findPairs}
\alias{findPairs}
\title{findPairs}
\usage{
findPairs(ds, LL, UL)
}
\arguments{
\item{ds}{The data set with the possible peaks}

\item{LL}{The lower limit multiplier for the range of the G2 peak, i.e, LL*G1}

\item{UL}{The upper limit multiplier for the range of the G2 peak, i.e, UL*G1}
}
\description{
findPairs identifies G1 and G2 sub population pairing.
}
\examples{
findPairs(
ds = peaks_ds,
LL = 1.8,
UL = 2.1
)
}
