% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputData.R
\name{outputData}
\alias{outputData}
\title{outputData}
\usage{
outputData(
  flowDir,
  singleDs,
  finishedDs,
  investigateDs,
  xVariable,
  doubletFlag,
  saveGraph
)
}
\arguments{
\item{flowDir}{The data set to analyze for doublets}

\item{singleDs}{The data set for the single populations found in PeakAlgorithm1}

\item{finishedDs}{The data set for all other flow frames that have been analyzed}

\item{investigateDs}{The data set that got identified as investigate}

\item{xVariable}{The fluorescence channel on the x axis}

\item{doubletFlag}{TRUE/FALSE}

\item{saveGraph}{TRUE/FALSE}
}
\description{
outputData creates a .csv and saves it in a folder called ‘analysis’ for the user.
The dataset includes the location of all peaks identified, the height of the peaks,
an option to have the information on doublets, a flag if the data was investigate,
a confidence interval for the population.
}
\examples{
outputData(
 flowDir = "FlowData/T10_FLC/gated_data",
 singleDs = single_data,
 finishedDs = finished_data,
 investigateDs = investigate_data,
 xVariable = "FITC-A",
 doubletFlag = FALSE,
 saveGraph = TRUE
 )
}
