% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakAlgorithm1.R
\name{peakAlgorithm1}
\alias{peakAlgorithm1}
\title{peakAlgorithm1}
\usage{
peakAlgorithm1(flowDir, flowSet, xVariable, singleThreshold = 8)
}
\arguments{
\item{flowDir}{The directory of the gated .fcs data}

\item{flowSet}{List of names of the flow frames in the flow set}

\item{xVariable}{The fluorescence channel on the x axis}

\item{singleThreshold}{threshold for classifying single populations}
}
\description{
peakAlgorithm1 is the first branching point to identify samples with single
populations (which contain only two peaks and can be easily analysed) and to
flag everything else. The algorithm will find a single G1 and G2 pairing.
With that population, a boundary is set to the right of the G2 peak and the
ratio of cells in that area are calculated. If this ratio is within a certain
threshold, the population is marked as a single population. If the ratio exceeds
the threshold, the population is flagged to pass to the next algorithm, PeakAlgorithm2.
}
\examples{
peakAlgorithm1(
 flowDir = "FlowData/T10_FLC/gated_data",
 flowSet = flow_names_ds ,
 xVariable = "FITC-A"
)

}
