% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakAlgorithm3.R
\name{peakAlgorithm3}
\alias{peakAlgorithm3}
\title{peakAlgorithm3}
\usage{
peakAlgorithm3(
  flowDir,
  flaggedData_,
  xVariable,
  appendData,
  usedCellsThreshold,
  maxDoubletHeight
)
}
\arguments{
\item{flowDir}{The directory of the gated .fcs data}

\item{flaggedData_}{List of names of the flow frames that got flagged by PeakAlgorithm2}

\item{xVariable}{The fluorescence channel on the x axis}

\item{appendData}{The data set of flow frames that are finished being analyzed
from the previous peak algorithms. The flow frames that are done after peakAlgorithm3 will append to this data set.}

\item{usedCellsThreshold}{Threshold for classifying multiple populations}

\item{maxDoubletHeightThe}{maximum height a doublet can be. If left as NA
the algorithm will find a value based on the other peaks}
}
\description{
The third branching point is in peakAlgorithm3, which seeks to identify all
possible sub populations that got flagged from PeakAlgorithm2. The algorithm
looks at the proportion of cells used in the subpopulations (CellProp).
In other words, once the populations have been identified for a given flow frame,
peakAlgorithm3 will find the cells associated with the G1/G2 pairs. The difference
from PeakAlgorithm2 is that peakAlgorithm3 will try and find a G2 pair in a (1.7\emph{G1, 2.3}G1)
}
\examples{
\dontrun{
peakAlgorithm3(
 flowDir = "FlowData/T10_FLC/gated_data",
 flaggedData_ = flaggedData ,
 xVariable = "FITC-A",
 appendData = analyzedDs,
 usedCellsThreshold = 86,
 maxDoubletHeight = 50
 )
 }

}
