% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakAlgorithm4.R
\name{peakAlgorithm4}
\alias{peakAlgorithm4}
\title{peakAlgorithm4}
\usage{
peakAlgorithm4(
  flowDir,
  flaggedData_,
  xVariable,
  appendData,
  usedCellsThreshold,
  MaxDoubletHeight
)
}
\arguments{
\item{flowDir}{The directory of the gated .fcs data}

\item{flaggedData_}{List of names of the flow frames that got flagged by PeakAlgorithm3}

\item{xVariable}{The fluorescence channel on the x axis}

\item{appendData}{The data set of flow frames that are finished being analyzed
from the previous peak algorithms. The flow frames that are done after peakAlgorithm4
will append to this data set.}

\item{usedCellsThreshold}{Threshold for classifying multiple populations}

\item{MaxDoubletHeight}{}
}
\description{
The forth branching point is in peakAlgorithm4, which seeks to identify all
possible sub populations that got flagged from PeakAlgorithm3. The algorithm
looks at the proportion of cells used in the subpopulations (CellProp).
In other words, once the populations have been identified for a given flow frame,
peakAlgorithm4 will find the cells associated with the G1/G2 pairs. The difference
from PeakAlgorithm3 is that peakAlgorithm4 works with more granular data, i.e, smooth level = 4.
Additionally, those that get flagged from peakAlgorithm4 will get a 'messy' indicator.
Letting the user know that they may want to manually look those who got flagged as messy.
}
\examples{
peakAlgorithm4(
 flowDir = "FlowData/T10_FLC/gated_data",
 flaggedData_ = flagged_data ,
 xVariable = "FITC-A",
 appendData = analyzed_ds,
 usedCellsThreshold = 86,
 MaxDoubletHeight = 50
 )

}
