% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakAlgorithm5.R
\name{peakAlgorithm5}
\alias{peakAlgorithm5}
\title{peakAlgorithm5}
\usage{
peakAlgorithm5(flowDir, flaggedData_, xVariable, appendData)
}
\arguments{
\item{flowDir}{The directory of the gated .fcs data}

\item{flaggedData_}{List of names of the flow frames that got flagged by PeakAlgorithm3}

\item{xVariable}{The fluorescence channel on the x axis}

\item{appendData}{The data set of flow frames that are finished being analyzed
from the previous peak algorithms.}
}
\description{
The last branching point is in peakAlgorithm5, which seeks to identify all
possible sub populations that got flagged from PeakAlgorithm4. The algorithm
looks at the proportion of cells used in the subpopulations (CellProp).
In other words, once the populations have been identified for a given flow frame,
peakAlgorithm5 will find the cells associated with the G1/G2 pairs. The difference
from PeakAlgorithm4 is that peakAlgorithm5 works with more granular data, i.e, smooth level = 3, as well try and find a G2 pair in a (1.5\emph{G1, 2.5}G1)
}
\examples{
\dontrun{
peakAlgorithm5(
 flowDir = "FlowData/T10_FLC/gated_data",
 flaggedData_ = flagged_data ,
 xVariable = "FITC-A",
 appendData = analyzed_ds
 )
 }

}
