% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popConfidence2Pop.R
\name{popConfidence2Pop}
\alias{popConfidence2Pop}
\title{popConfidence2Pop}
\usage{
popConfidence2Pop(flowDir, ds, xVariable, saveGraph = TRUE)
}
\arguments{
\item{flowDir}{The directory of the gated .fcs data}

\item{ds}{dataset to be analyzed}

\item{xVariable}{The fluorescence channel on the x axis}

\item{saveGraph}{T/F for saving the graphs as an output of the NLS}
}
\description{
popConfidence2Pop
}
\examples{
popConfidence2Pop(
 flowDir = "FlowData/T10_FLC/gated_data",
 ds = data,
 xVariable = "FITC-A",
 saveGraph = TRUE
 )


}
