% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popConfidence3Pop.R
\name{popConfidence3Pop}
\alias{popConfidence3Pop}
\title{popConfidence3Pop}
\usage{
popConfidence3Pop(flowDir, ds, xVariable, saveGraph = TRUE)
}
\arguments{
\item{flowDir}{The directory of the gated .fcs data}

\item{ds}{dataset to be analyzed}

\item{xVariable}{The fluorescence channel on the x axis}

\item{saveGraph}{T/F for saving the graphs as an output of the NLS}
}
\description{
popConfidence3Pop
}
\examples{
\dontrun{
popConfidence3Pop(
 flowDir = "FlowData/T10_FLC/gated_data",
 ds = data,
 xVariable = "FITC-A",
 saveGraph = TRUE
 )
}

}
