% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popConfidenceDoublet.R
\name{popConfidenceDoublet}
\alias{popConfidenceDoublet}
\title{popConfidenceDoublet}
\usage{
popConfidenceDoublet(flowDir, ds, xVariable, saveGraph = TRUE)
}
\arguments{
\item{flowDir}{The directory of the gated .fcs data}

\item{ds}{dataset to be analyzed}

\item{xVariable}{The fluorescence channel on the x axis}

\item{saveGraph}{T/F for saving the graphs as an output of the NLS}
}
\description{
popConfidenceDoublet
}
\examples{
\dontrun{
popConfidenceDoublet(
 flowDir = "FlowData/T10_FLC/gated_data",
 ds = data,
 xVariable = "FITC-A",
 saveGraph = TRUE
 )
}

}
