% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popConfidenceInitial.R
\name{popConfidenceInitial}
\alias{popConfidenceInitial}
\title{popConfidenceInitial}
\usage{
popConfidenceInitial(flowDir, ds, xVariable, saveGraph = TRUE)
}
\arguments{
\item{flowDir}{The directory of the gated .fcs data}

\item{ds}{dataset to be analyzed}

\item{saveGraph}{T/F for saving the graphs as an output of the NLS}

\item{x_variable}{The fluorescence channel on the x axis}
}
\description{
popConfidenceInitial
}
\examples{
popConfidenceInitial(
 flowDir = "FlowData/T10_FLC/gated_data",
 ds = data,
 xVariable = "FITC-A",
 saveGraph = TRUE
 )


}
