% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSEOutlierDetection.R
\name{RSEOutlierDetection}
\alias{RSEOutlierDetection}
\title{RSEOutlierDetection}
\usage{
RSEOutlierDetection(
  xVariable = "FL1-A",
  flowDir = NA,
  filePath = NA,
  fileName = "RSEOutlier",
  alpha = 0.05
)
}
\arguments{
\item{xVariable}{The fluorescence channel on the x-axis (determined by the
set-up of your flow cytometer; ex. FLA-1, FITC-A)}

\item{flowDir}{The directory of the gated .fcs files. If NA, then a window
will prompt the user to select a folder (string file path)}

\item{filePath}{The directory of the output .csv file from
\code{flowPeakDetection()} that corresponds to the gated data (string file path)}

\item{fileName}{To label the PDF file of plotted outlier samples (string)}

\item{alpha}{The top percentage used to identify outliers (i.e 0.05)}
}
\value{
a .csv with an added column for outlier RSE about each sample, a
histogram to show the distribution of RSE values, and a PDF file of plotted
samples that were identified as outliers. Note: only the histogram is saved
if the skewness is less than zero.
}
\description{
A function that permits the user to analyze the RSE values provided by the
output from \code{flowPeakDetection()}. A new column is added to the .csv file
that indicates whether each sample is an outlier based on its RSE. The alpha
parameter allows the user to specify the quantile percentage they'd like to
use as a cutoff. Each sample that is an outlier is plotted in a grid
formation using \code{flowLineGraph()} where the dots on each plot are the G1
(blue) and G2 (green) peaks with the respective RSE in the top right corner.
}
\examples{
library(BiocFileCache)
original_wd <- getwd()
cachePath <- tempfile(pattern = "bioc_cache_")
dir.create(cachePath, showWarnings = FALSE)
bfc <- BiocFileCache(cachePath, ask = FALSE)
on.exit({
 if(normalizePath(cachePath) == normalizePath(getwd())){
   setwd(tempdir())  
 }
 unlink(cachePath, recursive = TRUE, force = TRUE)
}, add = TRUE)
gatedFolder <- file.path(cachePath, "fcs_data")
dir.create(gatedFolder, showWarnings = FALSE)
urls <- c(
  "https://ploidypeaksvignette.blob.core.windows.net/ploidypeaksvignettedata/gated_data/A1_1.fcs",
  "https://ploidypeaksvignette.blob.core.windows.net/ploidypeaksvignettedata/gated_data/A1_2.fcs",
  "https://ploidypeaksvignette.blob.core.windows.net/ploidypeaksvignettedata/gated_data/A1_3.fcs",
  "https://ploidypeaksvignette.blob.core.windows.net/ploidypeaksvignettedata/gated_data/A1_4.fcs",
  "https://ploidypeaksvignette.blob.core.windows.net/ploidypeaksvignettedata/gated_data/A1_5.fcs",
  "https://ploidypeaksvignette.blob.core.windows.net/ploidypeaksvignettedata/gated_data/A1_6.fcs"
)
rids <- vapply(urls, function(url){
  res <- bfcquery(bfc, url)
  if(nrow(res) == 0){
    bfcadd(bfc, basename(url), url)
  }else{
    res$rid[1]
  }
}, character(1))
fpaths <- bfcrpath(bfc, urls)
newFPaths <- file.path(gatedFolder, basename(urls))
file.copy(fpaths, newFPaths, overwrite = TRUE)
downloadedFiles <- newFPaths
csvFolder <- file.path(cachePath, "csv_data")
dir.create(csvFolder, showWarnings = FALSE)
csv <- "https://ploidypeaksvignette.blob.core.windows.net/ploidypeaksvignettedata/analysis/ploidyPeaksOutput.csv"
rid <- bfcquery(bfc, csv, field = "rname")$rid
if (length(rid) == 0) {
  rid <- bfcadd(bfc, csv, csv)
}
cachedFile <- bfcrpath(bfc, rid)
analysisFile <- file.path(csvFolder, basename(csv))
if (!file.exists(analysisFile)) file.copy(cachedFile, analysisFile)
RSEOutlierDetection(
 xVariable = "FITC-A",
 flowDir = gatedFolder,
 filePath = analysisFile,
 fileName = "gatedDataRSE",
 alpha = 0.05
)
}
