% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothData.R
\name{smoothData}
\alias{smoothData}
\title{smoothData}
\usage{
smoothData(flowDs, xVariable, smoothLevel)
}
\arguments{
\item{flowDs}{The directory where the data is located. If NA then a window
will prompt the user to select a folder.}

\item{xVariable}{The fluorescence channel you are interested in analyzing}

\item{smoothLevel}{The level of smoothing applied to the data}
}
\description{
A function that reads in the gated data and applies a smoothing factor
to the data.
}
\examples{
\dontrun{
smoothData(flowDs, "FITC-A", 5)
}
}
