% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatedMeans.R
\name{updatedMeans}
\alias{updatedMeans}
\title{updatedMeans}
\usage{
updatedMeans(ds, flowDir, xVariable)
}
\arguments{
\item{ds}{The dataset that has a smoothing factor of 13.}

\item{flowDir}{The directory of the gated .fcs data}

\item{xVariable}{The fluorescence channel on the x axis}
}
\description{
A function that will update the peaks identified for the single populations
to find their true peaks. The single populations are smoothed out by a factor
of 13 and so updatedMeans will find the peaks associated with the data at
a smoothing factor of 5.
}
\examples{
\dontrun{
updatedMeans(
ds = singleData,
flowDir = "FlowData/T10_FLC/gated_data",
xVariable = "FITC-A"
)
}
}
