% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowLineGraph.R
\name{flowLineGraph}
\alias{flowLineGraph}
\title{flowLineGraph}
\usage{
flowLineGraph(
  flowDir = NA,
  flowControl = NA,
  flowSamples = NA,
  xVariable,
  flowColours = NA
)
}
\arguments{
\item{flowDir}{The directory where the data is located. If NA then a window
will prompt the user to select a folder.}

\item{flowControl}{The control sample}

\item{flowSamples}{The vector of samples}

\item{xVariable}{The fluorescence channel you are interested in visualizing}

\item{flowColours}{Vector of colours for the flowSample to be plotted}
}
\value{
a line graph for the samples
}
\description{
A function that permits the users to plot as many flow frames as they would
like on the same graph panel. Each of the colourful lines represent a
different flow frame, and the black line represents the control sample used
for comparison (control sample is indicated by the user).
}
\examples{
 flowLineGraph(
  flowControl = "A01-A01",
  flowSamples = c("A07-G12","A13-E06", "A02-B10", "T1-D08"),
  flowColours = c("green","blue", "red", "orange"),
  xVariable = "FITC-A",
  flowDir = paste0(system.file(package = "PloidyPeaks"), "/gated_data/")
 )
}
