% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowLineGraph.R
\name{flowLineGraph}
\alias{flowLineGraph}
\title{flowLineGraph}
\usage{
flowLineGraph(
  flowDir = NA,
  flowControl = NA,
  flowSamples = NA,
  xVariable = "FL1-A",
  flowColours = NA,
  grid = FALSE,
  annotations = NA,
  vertLine1 = NA,
  vertLine2 = NA,
  fileName = "Samples",
  samplePeaks = NA
)
}
\arguments{
\item{flowDir}{The directory where the data is located. If NA, then a window
will prompt the user to select a folder.}

\item{flowControl}{The control sample}

\item{flowSamples}{The vector of samples}

\item{xVariable}{The fluorescence channel you are interested in visualizing}

\item{flowColours}{Vector of colours for the flowSample to be plotted (only
used if grid = FALSE)}

\item{grid}{Whether you want to plot samples individual samples in a grid or
all together on one graph panel}

\item{annotations}{The vector of annotations to plot on each graph in the
grid where the first entry is the label for the annotation}

\item{vertLine1}{The vector of values for a vertical line to be plotted on
each graph in the grid}

\item{vertLine2}{The vector of values for a second vertical line to be
plotted on each graph in the grid}

\item{fileName}{To label the PDF file if plotting as a grid}

\item{samplePeaks}{The matrix of G1 and G2 peak values for each sample (NOTE:
this is only used in the \code{RSEOutlierDetection()} function)}
}
\value{
either a single line graph for the samples or a PDF file of samples
plotted in a grid
}
\description{
A function that permits the users to plot multiple flow frames, either all
on the same graph panel or separately on a grid. Each colourful line on the
same panel represents a different flow frame, and the black line represents
the control sample used for comparison (if indicated by the user). If plotted
on a grid, each flow frame is the black line, with the optional control
sample as the grey line. Annotations as well as vertical lines can be added
to each graph in the grid. If the user chooses to only plot a control sample,
they can also specify annotations and vertical lines, but the 'grid'
parameter will be ignored.
}
\examples{
 flowLineGraph(
  flowControl = "A1_1.fcs",
  flowSamples = c("A1_2.fcs","A1_3.fcs", "A1_4.fcs", "A1_5.fcs", "A1_6.fcs"),
  flowColours = c("green", "blue", "red", "orange", "purple"),
  xVariable = "FITC-A",
  flowDir = dirname(cachedGatedData[[1]]),
  grid = FALSE
 )
 
 flowLineGraph(
  flowControl = "A1_1.fcs",
  flowSamples = c("A1_2.fcs","A1_3.fcs", "A1_4.fcs", "A1_5.fcs", "A1_6.fcs"),
  flowColours = NA,
  xVariable = "FITC-A",
  flowDir = dirname(cachedGatedData[[1]]),
  grid = TRUE,
  annotations = c("RSE", 12.41, 13.16, 13.45, 8.15, 4.21),
  vertLine1 = c(150, 155, 150, 145, 140),
  vertLine2 = c(335, 330, 340, 330, 305),
  fileName = "Gated_Grid"
  )
}
