% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massFlowGraph.R
\name{massFlowGraph}
\alias{massFlowGraph}
\title{massFlowGraph}
\usage{
massFlowGraph(xVariable = "FITC-A", flowDir = NA, filePath = NA, fileName = NA)
}
\arguments{
\item{xVariable}{The fluorescence channel on the x-axis}

\item{flowDir}{The directory of the gated .fcs data}

\item{filePath}{The directory of the output .csv file from
\code{flowPeakDetection()} that corresponds to the gated data}

\item{fileName}{To label the PDF file produced internally by \code{flowLineGraph()}}
}
\value{
description
}
\description{
A function that permits the user to plot flow frames using the output from
\code{flowPeakDetection()}. The user is prompted to choose which set of frames to
plot (all, flagged, or doublet flagged). The chosen flow frames are plotted
using the grid functionality in \code{flowLineGraph()}. The user must ensure the
flow directory and file path correspond to the same set of samples for plotting.
}
\examples{
massFlowGraph(
 xVariable = "FITC-A",
 flowDir = paste0(system.file(package = "PloidyPeaks"), "/gated_data/"),
 filePath = paste0(system.file(package = "PloidyPeaks"), "/analysis/"),
 fileName = "gatedDataRSE"
 )
}
