% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massFlowGraph.R
\name{massFlowGraph}
\alias{massFlowGraph}
\title{massFlowGraph}
\usage{
massFlowGraph(xVariable = "FITC-A", flowDir = NA, filePath = NA, fileName = NA)
}
\arguments{
\item{xVariable}{The fluorescence channel on the x-axis}

\item{flowDir}{The directory of the gated .fcs data}

\item{filePath}{The directory of the output .csv file from
\code{flowPeakDetection()} that corresponds to the gated data}

\item{fileName}{To label the PDF file produced internally by \code{flowLineGraph()}}
}
\value{
description
}
\description{
A function that permits the user to plot flow frames using the output from
\code{flowPeakDetection()}. The user is prompted to choose which set of frames to
plot (all, flagged, or doublet flagged). The chosen flow frames are plotted
using the grid functionality in \code{flowLineGraph()}. The user must ensure the
flow directory and file path correspond to the same set of samples for plotting.
}
\examples{
library(BiocFileCache)
bfc <- BiocFileCache()
urls <- c(
  "https://ploidypeaksvignette.blob.core.windows.net/ploidypeaksvignettedata/gated_data/A1_1.fcs",
  "https://ploidypeaksvignette.blob.core.windows.net/ploidypeaksvignettedata/gated_data/A1_2.fcs",
  "https://ploidypeaksvignette.blob.core.windows.net/ploidypeaksvignettedata/gated_data/A1_3.fcs",
  "https://ploidypeaksvignette.blob.core.windows.net/ploidypeaksvignettedata/gated_data/A1_4.fcs",
  "https://ploidypeaksvignette.blob.core.windows.net/ploidypeaksvignettedata/gated_data/A1_5.fcs",
  "https://ploidypeaksvignette.blob.core.windows.net/ploidypeaksvignettedata/gated_data/A1_6.fcs"
)
folderPath <- file.path(bfc@cache, "gated_data")
dir.create(folder_path, showWarnings = FALSE, recursive = TRUE)
cachedFiles <- sapply(urls, function(url) {
  cachedFile <- bfcadd(bfc, url)
  targetFile <- file.path(folderPath, basename(url))
  if (!file.exists(targetFile)) file.copy(cachedFile, targetFile)
})
subDir <- file.path(bfc@cache, "csv_data")
dir.create(subDir, showWarnings = FALSE, recursive = TRUE)
analysisFile <- file.path(subDir, basename(url))
if (!file.exists(analysisFile)) file.copy(bfcadd(bfc, url), analysisFile)
massFlowGraph(
  xVariable = "FITC-A",
  flowDir = paste0(dirname(cachedFiles[[1]]), "/gated_data"),
  filePath = analysisFile,
  fileName = "gatedDataRSE"
 )
}
