% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakCorrection.R
\name{peakCorrection}
\alias{peakCorrection}
\title{peakCorrection}
\usage{
peakCorrection(xVariable, flowDir = NA, sampleName, numSubPop = 2)
}
\arguments{
\item{xVariable}{The fluorescence channel on the x axis}

\item{flowDir}{The directory of the gated .fcs data}

\item{sampleName}{the name of the sample you want to analyze}

\item{numSubPop}{the number of subpopulations you want the algorithm to identify}
}
\description{
peakCorrection is a function created for the user to rerun the analysis on
a sample with a certain number of subpopulations.
}
\examples{
\dontrun{
peakCorrection(
 xVariable = "FITC-A",
 flowDir = NA,
 sampleName = "A01-A01",
 numSubPop = 2
 )
 }

}
