% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakCorrection.R
\name{peakCorrection}
\alias{peakCorrection}
\title{peakCorrection}
\usage{
peakCorrection(xVariable, flowDir = NA, sampleName, numSubPop = 2)
}
\arguments{
\item{xVariable}{The fluorescence channel on the x axis}

\item{flowDir}{The directory of the gated .fcs data}

\item{sampleName}{the name of the sample the user wants to analyze}

\item{numSubPop}{the number of subpopulations the user wants
the algorithm to identify}
}
\value{
a .csv with information about each sample and nls graphs
}
\description{
peakCorrection is a function created for the user to rerun the analysis on
a sample with a certain number of subpopulations.
}
\examples{
peakCorrection(
 xVariable = "FITC-A",
 flowDir = paste0(system.file(package = "PloidyPeaks"), "/gated_data/"),
 sampleName = "A02-B10",
 numSubPop = 2
 )

}
