% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakCorrection.R
\name{peakCorrection}
\alias{peakCorrection}
\title{peakCorrection}
\usage{
peakCorrection(xVariable = "FL1-A", flowDir = NA, sampleName, numSubPop = 2)
}
\arguments{
\item{xVariable}{The fluorescence channel on the x axis}

\item{flowDir}{The directory of the gated .fcs data}

\item{sampleName}{the name of the sample the user wants to analyze}

\item{numSubPop}{the number of subpopulations the user wants
the algorithm to identify

@import flowTime}
}
\value{
a .csv with information about each sample and nls graphs
}
\description{
peakCorrection is a function created for the user to rerun the analysis on
a sample with a certain number of subpopulations.
}
\examples{
library(BiocFileCache)
bfc <- BiocFileCache()
urls <- c(
  "https://ploidypeaksvignette.blob.core.windows.net/ploidypeaksvignettedata/gated_data/A1_1.fcs",
  "https://ploidypeaksvignette.blob.core.windows.net/ploidypeaksvignettedata/gated_data/A1_2.fcs",
  "https://ploidypeaksvignette.blob.core.windows.net/ploidypeaksvignettedata/gated_data/A1_3.fcs",
  "https://ploidypeaksvignette.blob.core.windows.net/ploidypeaksvignettedata/gated_data/A1_4.fcs",
  "https://ploidypeaksvignette.blob.core.windows.net/ploidypeaksvignettedata/gated_data/A1_5.fcs",
  "https://ploidypeaksvignette.blob.core.windows.net/ploidypeaksvignettedata/gated_data/A1_6.fcs"
)
folderPath <- file.path(bfc@cache, "gated_data")
dir.create(folderPath, showWarnings = FALSE, recursive = TRUE)
cachedFiles <- sapply(urls, function(url) {
  cachedFile <- bfcadd(bfc, url)
  targetFile <- file.path(folderPath, basename(url))
  if (!file.exists(targetFile)) file.copy(cachedFile, targetFile)
})
peakCorrection(
 xVariable = "FITC-A",
 flowDir = paste0(dirname(cachedFiles[[1]]), "/gated_data"),
 sampleName = "A1_4.fcs",
 numSubPop = 2
 )

}
