% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectGateFlowFrame.R
\name{rectGateFlowFrame}
\alias{rectGateFlowFrame}
\title{rectGateFlowFrame}
\usage{
rectGateFlowFrame(
  rawDir = NA,
  flowName,
  xVariable = "FL1-A",
  yVariable = "SSC-A",
  xMinValue = 10000,
  xMaxValue = 9e+05,
  yMinValue = 10000,
  yMaxValue = 9e+05,
  savePlot = TRUE
)
}
\arguments{
\item{rawDir}{The directory of the raw .fcs data}

\item{flowName}{The name of the .fcs file}

\item{xVariable}{The fluorescence channel on the x axis - by default "FL1-A"}

\item{yVariable}{The fluorescence channel on the y axis - by default "SSC-A"}

\item{xMinValue}{The lower bound x value for the gate - by default 10000}

\item{xMaxValue}{The upper bound x value for the gate - by default 900000}

\item{yMinValue}{The lower bound y value for the gate - by default 10000}

\item{yMaxValue}{The upper bound y value for the gate - by default 900000}

\item{savePlot}{A side by side graph comparison the raw data and the gated
data - by default TRUE}
}
\value{
A .fcs of the gated data
}
\description{
This function will gate your raw .fcs file and save your gated data, a plot
of your gated data, and a plot of the original data
}
\examples{
rectGateFlowFrame(
 rawDir = NA,
 flowName = "A10-t0.fcs",
 xVariable = "FL1-A",
 yVariable = "SSC-A",
 xMinValue = 10000,
 xMaxValue = 900000,
 yMinValue = 10000,
 yMaxValue = 900000,
 savePlot = TRUE
)
}
