% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectGateFlowSet.R
\name{rectGateFlowSet}
\alias{rectGateFlowSet}
\title{rectGateFlowSet}
\usage{
rectGateFlowSet(
  rawDir = NA,
  xVariable = "FL1-A",
  yVariable = "SSC-A",
  xMinValue = 10000,
  xMaxValue = 9e+05,
  yMinValue = 10000,
  yMaxValue = 9e+05,
  savePlot = TRUE
)
}
\arguments{
\item{rawDir}{The directory of the raw .fcs data}

\item{xVariable}{The fluorescence channel on the x axis}

\item{yVariable}{The fluorescence channel on the y axis}

\item{xMinValue}{The lower bound x value for the gate}

\item{xMaxValue}{The upper bound x value for the gate}

\item{yMinValue}{The lower bound y value for the gate}

\item{yMaxValue}{The upper bound y value for the gate}

\item{savePlot}{A side by side graph comparison the raw data and the gated
data - by default TRUE}
}
\value{
A set of .fcs of the gated data, plots of gated data, and a .csv
containing information on how percentage of cells gated out
}
\description{
This function will gate your raw flow set and save your gated data, a plot
of your gated data, and a plot of the original data
}
\examples{
\donttest{
rectGateFlowSet(
 rawDir = paste0(system.file(package = "PloidyPeaks"), "/raw_data"),
 xVariable = "FITC-A",
 yVariable = "SSC-A",
 xMinValue = 50,
 xMaxValue = 850,
 yMinValue = 50,
 yMaxValue = 850,
 savePlot = FALSE
)
}
}
