% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectGateFlowSet.R
\name{rectGateFlowSet}
\alias{rectGateFlowSet}
\title{rectGateFlowSet}
\usage{
rectGateFlowSet(
  raw_dir = NA,
  x_variable = "FL1-A",
  y_variable = "SSC-A",
  x_min_value = 10000,
  x_max_value = 9e+05,
  y_min_value = 10000,
  y_max_value = 9e+05,
  save_plot = TRUE
)
}
\arguments{
\item{raw_dir}{The directory of the raw .fcs data}

\item{x_variable}{The fluorescence channel on the x axis - by default "FL1-A"}

\item{y_variable}{The fluorescence channel on the y axis - by default "SSC-A"}

\item{x_min_value}{The lower bound x value for the gate - by default 10000}

\item{x_max_value}{The upper bound x value for the gate - by default 900000}

\item{y_min_value}{The lower bound y value for the gate - by default 10000}

\item{y_max_value}{The upper bound y value for the gate - by default 900000}

\item{save_plot}{A side by side graph comparison the raw data and the gated
data - by default TRUE}
}
\value{
A .fcs of the gated data
}
\description{
This function will gate your raw flow set and save your gated data, a plot
of your gated data, and a plot of the original data
}
\examples{
rectGateFlowSet(
 raw_dir = NA,
 x_variable = "FL1-A",
 y_variable = "SSC-A",
 x_min_value = 10000,
 x_max_value = 900000,
 y_min_value = 10000,
 y_max_value = 900000,
 save_plot = TRUE
)
}
