% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectGateFlowSet.R
\name{rectGateFlowSet}
\alias{rectGateFlowSet}
\title{rectGateFlowSet}
\usage{
rectGateFlowSet(
  rawDir = NA,
  xVariable = "FL1-A",
  yVariable = "SSC-A",
  xMinValue = 10000,
  xMaxValue = 9e+05,
  yMinValue = 10000,
  yMaxValue = 9e+05,
  savePlot = TRUE
)
}
\arguments{
\item{rawDir}{The directory of the raw .fcs data}

\item{xVariable}{The fluorescence channel on the x axis}

\item{yVariable}{The fluorescence channel on the y axis}

\item{xMinValue}{The lower bound x value for the gate}

\item{xMaxValue}{The upper bound x value for the gate}

\item{yMinValue}{The lower bound y value for the gate}

\item{yMaxValue}{The upper bound y value for the gate}

\item{savePlot}{A side by side graph comparison the raw data and the gated
data - by default TRUE}
}
\value{
A set of .fcs of the gated data, plots of gated data, and a .csv
containing information on how percentage of cells gated out
}
\description{
This function will gate your raw flow set and save your gated data, a plot
of your gated data, and a plot of the original data
}
\examples{
library(BiocFileCache)
bfc <- BiocFileCache()
urls <- c(
  "https://ploidypeaksvignette.blob.core.windows.net/ploidypeaksvignettedata/raw_data/A1_1.fcs",
  "https://ploidypeaksvignette.blob.core.windows.net/ploidypeaksvignettedata/raw_data/A1_2.fcs",
  "https://ploidypeaksvignette.blob.core.windows.net/ploidypeaksvignettedata/raw_data/A1_3.fcs",
  "https://ploidypeaksvignette.blob.core.windows.net/ploidypeaksvignettedata/raw_data/A1_4.fcs",
  "https://ploidypeaksvignette.blob.core.windows.net/ploidypeaksvignettedata/raw_data/A1_5.fcs",
  "https://ploidypeaksvignette.blob.core.windows.net/ploidypeaksvignettedata/raw_data/A1_6.fcs"
)
folderPath <- file.path(bfc@cache, "raw_data")
dir.create(folderPath, showWarnings = FALSE, recursive = TRUE)
cachedFiles <- sapply(urls, function(url) {
  cachedFile <- bfcadd(bfc, url)
  targetFile <- file.path(folderPath, basename(url))
  if (!file.exists(targetFile)) file.copy(cachedFile, targetFile)
})
rectGateFlowSet(
 rawDir = paste0(dirname(cachedFiles[[1]]), "/raw_data"),
 xVariable = "FITC-A",
 yVariable = "SSC-A",
 xMinValue = 50,
 xMaxValue = 800,
 yMinValue = 50,
 yMaxValue = 800,
 savePlot = FALSE
)
}
