% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSEOutlierDetection.R
\name{RSEOutlierDetection}
\alias{RSEOutlierDetection}
\title{RSEOutlierDetection}
\usage{
RSEOutlierDetection(
  xVariable = "FL1-A",
  flowDir = NA,
  filePath = NA,
  fileName = "RSEOutlier",
  alpha = 0.05
)
}
\arguments{
\item{xVariable}{The fluorescence channel on the x axis}

\item{flowDir}{The directory of the gated .fcs data}

\item{filePath}{The csv analysis file that pairs with the gated .fcs data}

\item{fileName}{To label the PDF file of plotted outlier samples}

\item{alpha}{The top percentage used to identify outliers (i.e 0.05)}
}
\value{
a .csv with an added column for outlier RSE about each sample, a
histogram to show the distribution of RSE values, and a PDF file of plotted
samples that were identified as outliers. Note: only the histogram is saved
if the skewness is less than zero.
}
\description{
A function that permits the user to analyze the RSE values provided by the
output from \code{flowPeakDetection()}. A new column is added to the .csv file
that indicates whether each sample is an outlier based on its RSE. The alpha
parameter allows the user to specify the quantile percentage they'd like to
use as a cutoff. Each sample that is an outlier is plotted in a grid
formation using \code{flowLineGraph()} where the dots on each plot are the G1
(blue) and G2 (green) peaks with the respective RSE in the top right corner.
}
\examples{
RSEOutlierDetection(
 xVariable = "FITC-A",
 flowDir = paste0(system.file(package = "PloidyPeaks"), "/gated_data/"),
 filePath = paste0(system.file(package = "PloidyPeaks"), "/analysis/"),
 fileName = "gatedDataRSE",
 alpha = 0.05
)
}
