% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatedMeans.R
\name{updatedMeans}
\alias{updatedMeans}
\title{updatedMeans}
\usage{
updatedMeans(ds)
}
\arguments{
\item{flowDs}{The dataset that has a smoothing factor of 13.}
}
\description{
A function that will update the peaks identified for the single populations
to find their true peaks. The single populations are smoothed out by a factor
of 13 and so updatedMeans will find the peaks associated with the data at
a smoothing factor of 5.
}
\examples{
updatedMeans(ds = singleData)
}
